refactor: base 프래그먼트 적용
