+++ Doraboda/app/src/main/java/com/univ/doraboda/DI/PlayerModule.kt
@@ -10,7 +10,7 @@ import androidx.media3.common.util.UnstableApi
 import androidx.media3.exoplayer.ExoPlayer
 import androidx.media3.session.MediaSession
 import com.univ.doraboda.callback.CustomCallback
-import com.univ.doraboda.view.MainActivity
+import com.univ.doraboda.ui.MainActivity
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
