+++ Doraboda/app/src/main/java/com/univ/doraboda/ui/BaseFragment.kt
@@ -1,0 +1,34 @@
+package com.univ.doraboda.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
+
+abstract class BaseFragment<T : ViewDataBinding>(): Fragment() {
+    var _binding: T? = null
+    val binding: T
+        get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = DataBindingUtil.inflate(inflater, layoutId(), container, false)
+        layoutInit()
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    abstract fun layoutId(): Int
+
+    abstract fun layoutInit()
+}
