--- Doraboda/app/src/main/java/com/univ/doraboda/view/MainActivity.kt
@@ -1,23 +1,0 @@
-package com.univ.doraboda.view
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.setupWithNavController
-import com.univ.doraboda.R
-import com.univ.doraboda.databinding.ActivityMainBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-    lateinit var binding: ActivityMainBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val navHostFragment = supportFragmentManager.findFragmentById(R.id.homenavfragment) as NavHostFragment
-        val navController = navHostFragment.navController
-        binding.homeBottomNavigation.setupWithNavController(navController)
-    }
-}
