+++ Doraboda/app/src/main/res/layout/activity_add_emotion.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.AddEmotionActivity">
+    tools:context=".ui.AddEmotionActivity">
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/AddEmotionToolbar"
         android:layout_width="match_parent"
