+++ Doraboda/app/src/main/res/layout/activity_settings.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.SettingsActivity">
+    tools:context=".ui.SettingsActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
