+++ Doraboda/app/src/main/res/layout/fragment_add_emotion.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.AddEmotionFragment">
+    tools:context=".ui.AddEmotionFragment">
 <LinearLayout
     android:id="@+id/addEmotionLinearLayout"
     android:layout_width="match_parent"
